requirements.txt
requests~=2.31.0
python-telegram-bot~=20.6
python-telegram-bot[job-queue]~=20.6
firebase-admin~=6.2.0
tweepy~=4.14.0
opencv-python-headless~=4.8.1.78
urllib3~=2.0.7
numpy~=1.26.1
Pillow~=9.5.0
pytz~=2023.3.post1
beautifulsoup4~=4.12.2